with fix

+++ 0x1A-hash_tables/0-hash_table_create.c
@@ -8,7 +8,7 @@
  */
 hash_table_t *hash_table_create(unsigned long int size)
 {
-	unsigned int inc;
+	unsigned long int inc;
 	hash_table_t *hash;
 
 	hash = malloc(sizeof(hash_table_t));
